Resolves #35 Allow use of "is*" methods for boolean property getters without phaff
Any method can now be annotated as a property. if the name starts with get or is the default property name becomes the method name with that prefix stripped. Otherwise the default name is the method name.

Made it possible to test the ClassAnnotationDriver.
Added a bunch of toString methods to make life simpler when testing.