Fixes issue #41 : ghostly class level annotations.
Added a workaround for what's probably a PHP bug. By calling token_get_all()
on a dummy docblock comment we set a compiler global to something "safe."
This means that if the compiler does try to use the global it wont muck up our config.

Also fixes the Annotation name parser, which was rather broken and didn't work if you
fully qualified a name in an annotation comment.